FastHenry 3.0 with superconductive elements support - 2014/08/25
- This is the fasthenry-3.0-12Nov96 distribution with the changes
  highlighted in the README file
- Commit is based on Whitleley Research Inc. version dated 2014/08/25:
  o Major improvements in the computation speed of large problems
  o output to Matlab option should work now, on any platform
  o Imported the modified indexing of the Jimag/Jreal/Jmag files from
    FastHenry2 code from FastFieldSolvers.com

+++ src/zbuf/mulGlobal.h
@@ -133,7 +133,7 @@ extern long memMSC;
 	 "\nfastcap: out of memory in file `%s' at line %d\n",              \
 	       __FILE__, __LINE__);                                         \
        (void)fprintf(stderr, " (NULL pointer on %ld byte request)\n",        \
-		     (NUM)*sizeof(TYPE));                                   \
+		     (long)(NUM)*sizeof(TYPE));                                   \
        DUMPALLOCSIZ;                                                        \
        DUMPRSS;                                                             \
        (void)fflush(stderr);                                                \
@@ -170,7 +170,7 @@ extern long memMSC;
 	 "\nfastcap: out of memory in file `%s' at line %d\n",               \
 	       __FILE__, __LINE__);                                          \
        (void)fprintf(stderr, " (NULL pointer on %ld byte request)\n",         \
-		     (NUM)*sizeof(TYPE));                                    \
+		     (long)(NUM)*sizeof(TYPE));                                    \
        DUMPALLOCSIZ;                                                         \
        DUMPRSS;                                                              \
        (void)fflush(stderr);                                                 \
